pe/os/test_co16.c: Skip MPAM test if it isn't supported
The MPAM test was checking PPTT flags even if MPAM is not supported.
This caused the test to fail when it should be skipped.
This change follows the test description in the SBSA Test Scenario
document, Test 16 Page 10, where it check MPAM support first.
https://github.com/ARM-software/sbsa-acs/blob/master/docs/
Arm_SBSA_Architecture_Compliance_Test_Scenario.pdf

Signed-off-by: Nhi Pham <nhi@os.amperecomputing.com>